requirements.txt
Django==4.1.5
django-graphql-jwt==0.3.4
djongo==1.3.6
graphene-django==3.0.0
Pillow==9.4.0